Consolidate eval cases and deepen the harness architecture.
Unify case prompts and assertions in one catalog file, extract suite orchestration and result I/O into focused modules, and remove duplicate assertion helpers and stale tool-config references. Fix gitignore so src/results/ is tracked while /results/ artifacts stay ignored. Add the @/ import alias for src/.
