+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules/
-results/
+/results/
 journal/
 *.env
