+++ deno.json
@@ -1,6 +1,7 @@
 {
   "nodeModulesDir": "auto",
   "imports": {
+    "@/": "./src/",
     "@ai-sdk/google": "npm:@ai-sdk/google@^3.0.78",
     "@std/cli/parse-args": "jsr:@std/cli@^1.0.22/parse-args",
     "@comunica/query-sparql-rdfjs-lite": "npm:@comunica/query-sparql-rdfjs-lite@^5.2.2",
