+++ scripts/render-compare-report.ts
@@ -1,70 +1,12 @@
-import { join } from "@std/path";
+import {
+  buildTable,
+  formatDelta,
+  formatPercent,
+} from "../src/reporting/markdown-table.ts";
+import { findNewestCompareResultPath } from "../src/results/result-store.ts";
 import type { EvalAssertionPassRate, EvalCompareResult } from "../src/types.ts";
 
 const DEFAULT_RESULTS_DIRECTORY = "results";
-
-/** formatPercent renders a pass rate as a one-decimal percentage. */
-function formatPercent(passRate: number): string {
-  return `${(passRate * 100).toFixed(1)}%`;
-}
-
-/** formatDelta renders a pass-rate delta in percentage points. */
-function formatDelta(delta: number): string {
-  const sign = delta > 0 ? "+" : "";
-  return `${sign}${(delta * 100).toFixed(1)} pts`;
-}
-
-/** escapeTableCell protects Markdown tables from pipes and line breaks. */
-function escapeTableCell(value: unknown): string {
-  return String(value ?? "")
-    .replaceAll("|", "\\|")
-    .replaceAll("\r\n", "<br>")
-    .replaceAll("\n", "<br>")
-    .trim();
-}
-
-/** buildTable renders a GitHub-flavored Markdown table. */
-function buildTable(headers: string[], rows: string[][]): string {
-  const headerRow = `| ${headers.map(escapeTableCell).join(" | ")} |`;
-  const separatorRow = `| ${headers.map(() => ":---").join(" | ")} |`;
-  const bodyRows = rows.map((row) =>
-    `| ${row.map((cell) => escapeTableCell(cell)).join(" | ")} |`
-  );
-  return [headerRow, separatorRow, ...bodyRows].join("\n");
-}
-
-/** findCompareResultPath locates the newest comparison JSON artifact. */
-async function findCompareResultPath(
-  resultsDirectory: string,
-): Promise<string> {
-  const configuredPath = Deno.env.get("COMPARE_RESULT_PATH");
-  if (configuredPath) {
-    return configuredPath;
-  }
-
-  let newestPath: string | undefined;
-  let newestModifiedAt = 0;
-  for await (const entry of Deno.readDir(resultsDirectory)) {
-    if (
-      !entry.isFile || !entry.name.startsWith("compare-") ||
-      !entry.name.endsWith(".json")
-    ) {
-      continue;
-    }
-    const filePath = join(resultsDirectory, entry.name);
-    const fileInfo = await Deno.stat(filePath);
-    const modifiedAt = fileInfo.mtime?.getTime() ?? 0;
-    if (modifiedAt >= newestModifiedAt) {
-      newestModifiedAt = modifiedAt;
-      newestPath = filePath;
-    }
-  }
-
-  if (!newestPath) {
-    throw new Error(`No compare-*.json result found in ${resultsDirectory}`);
-  }
-  return newestPath;
-}
 
 /** collectAssertionRate returns a named assertion's rate for a case when present. */
 function collectAssertionRate(
@@ -78,8 +20,10 @@ function collectAssertionRate(
 
 /** renderCompareReport renders tool config comparison JSON as Markdown. */
 export function renderCompareReport(compareResult: EvalCompareResult): string {
-  const [baselineToolConfigId, ...candidateToolConfigIds] =
-    compareResult.toolConfigIds;
+  const baselineToolConfigId = compareResult.baselineToolConfigId;
+  const candidateToolConfigIds = compareResult.toolConfigIds.filter(
+    (toolConfigId) => toolConfigId !== baselineToolConfigId,
+  );
   const caseRows = compareResult.caseComparisons.map((comparison) => {
     const baselineRate = comparison.passRatesByToolConfig[baselineToolConfigId];
     const candidateCells = candidateToolConfigIds.flatMap((toolConfigId) => {
@@ -175,7 +119,10 @@ export function renderCompareReport(compareResult: EvalCompareResult): string {
 if (import.meta.main) {
   const resultsDirectory = Deno.env.get("RESULTS_DIRECTORY") ??
     DEFAULT_RESULTS_DIRECTORY;
-  const compareResultPath = await findCompareResultPath(resultsDirectory);
+  const compareResultPath = await findNewestCompareResultPath(resultsDirectory);
+  if (!compareResultPath) {
+    throw new Error(`No compare-*.json result found in ${resultsDirectory}`);
+  }
   const compareResult = JSON.parse(
     await Deno.readTextFile(compareResultPath),
   ) as EvalCompareResult;
