+++ tests/runner/run-eval-suite.test.ts
@@ -1,0 +1,133 @@
+import { assertEquals } from "@std/assert";
+import { evalCases } from "../../src/cases/index.ts";
+import {
+  aggregateEvalStats,
+  buildCompareResult,
+  synthesizeStatsFromSuite,
+} from "../../src/runner/run-eval-suite.ts";
+import type { EvalCaseResult, EvalSuiteResult } from "../../src/types.ts";
+
+/** createEvalCaseResult builds a minimal case result for suite aggregation tests. */
+function createEvalCaseResult(
+  overrides: Partial<EvalCaseResult> & Pick<EvalCaseResult, "id">,
+): EvalCaseResult {
+  return {
+    description: overrides.description ?? overrides.id,
+    prompt: overrides.prompt ?? "",
+    output: overrides.output ?? "",
+    runCompleted: overrides.runCompleted ?? true,
+    success: overrides.success ?? true,
+    metadata: {
+      providerId: "google",
+      modelId: "gemini-3.1-flash-lite",
+      stepCount: overrides.metadata?.stepCount ?? 1,
+      latencyMs: overrides.metadata?.latencyMs ?? 0,
+      trajectory: overrides.metadata?.trajectory ?? [],
+      ...overrides.metadata,
+    },
+    assertions: overrides.assertions ?? [{
+      name: "final-answer-correct",
+      pass: overrides.success ?? true,
+    }],
+    toolSequence: overrides.toolSequence ?? [],
+    ...overrides,
+  };
+}
+
+Deno.test("aggregateEvalStats computes case and assertion pass rates across trials", () => {
+  const selectedCases = evalCases.slice(0, 1);
+  const trialResults = [
+    [createEvalCaseResult({
+      id: selectedCases[0].id,
+      success: true,
+      assertions: [
+        { name: "final-answer-correct", pass: true },
+        { name: "step-count-bounded", pass: false },
+      ],
+    })],
+    [createEvalCaseResult({
+      id: selectedCases[0].id,
+      success: false,
+      assertions: [
+        { name: "final-answer-correct", pass: false },
+        { name: "step-count-bounded", pass: true },
+      ],
+    })],
+  ];
+
+  const statsResult = aggregateEvalStats(
+    selectedCases,
+    trialResults,
+    "google",
+    "gemini-3.1-flash-lite",
+    "baseline",
+    0.5,
+  );
+
+  assertEquals(statsResult.trialCount, 2);
+  assertEquals(statsResult.success, true);
+  assertEquals(statsResult.casePassRates[0].passCount, 1);
+  assertEquals(statsResult.casePassRates[0].passRate, 0.5);
+});
+
+Deno.test("synthesizeStatsFromSuite mirrors a single successful suite result", () => {
+  const suiteResult: EvalSuiteResult = {
+    providerId: "google",
+    modelId: "gemini-3.1-flash-lite",
+    toolConfigId: "baseline",
+    timestamp: "2026-05-21T00:00:00.000Z",
+    success: true,
+    results: [
+      createEvalCaseResult({
+        id: "happy-path-search-then-sparql",
+        success: true,
+        assertions: [{ name: "final-answer-correct", pass: true }],
+      }),
+    ],
+  };
+
+  const statsResult = synthesizeStatsFromSuite(suiteResult, 1);
+  assertEquals(statsResult.trialCount, 1);
+  assertEquals(statsResult.casePassRates[0].passRate, 1);
+});
+
+Deno.test("buildCompareResult records the first tool config as baseline", () => {
+  const selectedCases = evalCases.slice(0, 1);
+  const statsResults = [
+    aggregateEvalStats(
+      selectedCases,
+      [[createEvalCaseResult({
+        id: selectedCases[0].id,
+        success: true,
+        assertions: [{ name: "final-answer-correct", pass: true }],
+      })]],
+      "google",
+      "gemini-3.1-flash-lite",
+      "baseline",
+    ),
+    aggregateEvalStats(
+      selectedCases,
+      [[createEvalCaseResult({
+        id: selectedCases[0].id,
+        success: false,
+        assertions: [{ name: "final-answer-correct", pass: false }],
+      })]],
+      "google",
+      "gemini-3.1-flash-lite",
+      "strict-eval",
+    ),
+  ];
+
+  const compareResult = buildCompareResult(
+    selectedCases,
+    statsResults,
+    {
+      providerId: "google",
+      modelId: "gemini-3.1-flash-lite",
+      trialCount: 1,
+    },
+  );
+
+  assertEquals(compareResult.baselineToolConfigId, "baseline");
+  assertEquals(compareResult.toolConfigIds, ["baseline", "strict-eval"]);
+});
